# Optimizing some structured matrix multiply methods