--- a/stdlib/LinearAlgebra/src/bidiag.jl
+++ b/stdlib/LinearAlgebra/src/bidiag.jl
@@ -631,6 +631,15 @@ function *(A::SymTridiagonal, B::Diagonal)
     A_mul_B_td!(Tridiagonal(zeros(TS, size(A, 1)-1), zeros(TS, size(A, 1)), zeros(TS, size(A, 1)-1)), A, B)
 end
 
+function *(A::Bidiagonal, B::Bidiagonal)
+    TS = promote_op(matprod, eltype(A), eltype(B))
+    if A.uplo == B.uplo
+        A_mul_B_td!(similar(B, TS, size(B, 1), size(B, 2)), A, B)
+    else
+        A_mul_B_td!(Tridiagonal(zeros(TS, size(A, 1)-1), zeros(TS, size(A, 1)), zeros(TS, size(A, 1)-1)), A, B)
+    end
+end
+
 function *(A::AbstractMatrix, B::BiTriSym)
     TS = promote_op(matprod, eltype(A), eltype(B))
     A_mul_B_td!(similar(A, TS), A, B)