This PR addresses #28345 and #27176 as well as some missed optimizations from #28883.

A summary of the changes:

- Dense times Structured matrices output dense matrices (not `SparseMatrixCSC`)
- `Bidiagonal`x`Tridiagonal`, `SymTridiagonal`x`SymTridiagonal`, and others output `SparseMatrixCSC` (not Dense). These changes were mentioned in #28883, but apparently I missed them in that PR.

Below are some benchmarks as well as the input/output types:

-------------------------------------------------------------------------

Before:


| A              | B              | A*B             | Time             |
|----------------|----------------|-----------------|------------------|
| Array          | Diagonal       | Array           | 0.029855 seconds |
| Array          | Bidiagonal     | SparseMatrixCSC | 2.767608 seconds |
| Array          | Tridiagonal    | SparseMatrixCSC | 2.813555 seconds |
| Array          | SymTridiagonal | SparseMatrixCSC | 2.826780 seconds |
| Diagonal       | Array          | Array           | 0.030034 seconds |
| Diagonal       | Diagonal       | Diagonal        | 0.000007 seconds |
| Diagonal       | Bidiagonal     | SparseMatrixCSC | 0.053484 seconds |
| Diagonal       | Tridiagonal    | SparseMatrixCSC | 0.060171 seconds |
| Diagonal       | SymTridiagonal | SparseMatrixCSC | 0.059392 seconds |
| Bidiagonal     | Array          | Array           | 0.013068 seconds |
| Bidiagonal     | Diagonal       | SparseMatrixCSC | 0.052733 seconds |
| Bidiagonal     | Bidiagonal     | Array           | 0.344665 seconds |
| Bidiagonal     | Tridiagonal    | Array           | 0.345147 seconds |
| Bidiagonal     | SymTridiagonal | Array           | 0.344073 seconds |
| Tridiagonal    | Array          | Array           | 0.014226 seconds |
| Tridiagonal    | Diagonal       | SparseMatrixCSC | 0.057033 seconds |
| Tridiagonal    | Bidiagonal     | Array           | 0.346379 seconds |
| Tridiagonal    | Tridiagonal    | Array           | 0.345191 seconds |
| Tridiagonal    | SymTridiagonal | Array           | 0.347181 seconds |
| SymTridiagonal | Array          | Array           | 0.011845 seconds |
| SymTridiagonal | Diagonal       | SparseMatrixCSC | 0.057408 seconds |
| SymTridiagonal | Bidiagonal     | Array           | 0.349033 seconds |
| SymTridiagonal | Tridiagonal    | Array           | 0.347890 seconds |
| SymTridiagonal | SymTridiagonal | Array           | 0.348104 seconds |

```julia
julia> versioninfo()
Julia Version 1.1.1
Commit 55e36cc308 (2019-05-16 04:10 UTC)
Platform Info:
  OS: Linux (x86_64-linux-gnu)
  CPU: Intel(R) Core(TM) i5-8250U CPU @ 1.60GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-6.0.1 (ORCJIT, skylake)
```

----------------------------------------------------------------

This PR:

| A              | B              | A*B             | Time             |
|----------------|----------------|-----------------|------------------|
| Array          | Diagonal       | Array           | 0.013508 seconds |
| Array          | Bidiagonal     | Array           | 0.024095 seconds |
| Array          | Tridiagonal    | Array           | 0.007269 seconds |
| Array          | SymTridiagonal | Array           | 0.024687 seconds |
| Diagonal       | Array          | Array           | 0.011103 seconds |
| Diagonal       | Diagonal       | Diagonal        | 0.000009 seconds |
| Diagonal       | Bidiagonal     | Bidiagonal      | 0.000103 seconds |
| Diagonal       | Tridiagonal    | Tridiagonal     | 0.000106 seconds |
| Diagonal       | SymTridiagonal | Tridiagonal     | 0.000108 seconds |
| Bidiagonal     | Array          | Array           | 0.013104 seconds |
| Bidiagonal     | Diagonal       | Bidiagonal      | 0.000111 seconds |
| Bidiagonal     | Bidiagonal     | SparseMatrixCSC | 0.002295 seconds |
| Bidiagonal     | Tridiagonal    | SparseMatrixCSC | 0.002942 seconds |
| Bidiagonal     | SymTridiagonal | SparseMatrixCSC | 0.003073 seconds |
| Tridiagonal    | Array          | Array           | 0.022158 seconds |
| Tridiagonal    | Diagonal       | Tridiagonal     | 0.000108 seconds |
| Tridiagonal    | Bidiagonal     | SparseMatrixCSC | 0.002936 seconds |
| Tridiagonal    | Tridiagonal    | SparseMatrixCSC | 0.003766 seconds |
| Tridiagonal    | SymTridiagonal | SparseMatrixCSC | 0.003762 seconds |
| SymTridiagonal | Array          | Array           | 0.011779 seconds |
| SymTridiagonal | Diagonal       | Tridiagonal     | 0.000117 seconds |
| SymTridiagonal | Bidiagonal     | SparseMatrixCSC | 0.003116 seconds |
| SymTridiagonal | Tridiagonal    | SparseMatrixCSC | 0.003845 seconds |
| SymTridiagonal | SymTridiagonal | SparseMatrixCSC | 0.003791 seconds |