@mcognetta 2019-07-09
I resolved the ambiguities by restricting the left side to be `::Array`. I don't think this is a particularly good solution (since there could be other subtypes of `AbstractMatrix` that need consideration), but the alternative involved untangling a ton of ambiguous matrix multiply definitions. Suggestions/comments are appreciated.