@andreasnoack 2019-07-24
> I resolved the ambiguities by restricting the left side to be `::Matrix`

Did you try with `StridedMatrix`?