@mcognetta 2019-07-26
Fixed merge conflicts.