@mcognetta 2019-08-05
Seems like I accidentally deleted the relevant code. Should be fixed now.