@andreasnoack 2019-08-06
Thanks. One concern here is that `BidiagonalxTridiagonal` returning a sparse matrix implies a dependency of `LinearAlgebra` on `SparseArrays`. We already have a few of these so maybe we just want to accept it but we have been thinking about making the sparse stdlibs more stand-alone and pluggable in which case it would be unfortunate to add another such dependency. An alternative could be to add an informative error message for `BidiagonalxTridiagonal` asking the user to convert to sparse matrices before multiplying. I doubt that this will be a common operation so I'd expect few users to hit it.