@mcognetta 2019-08-06
@andreasnoack Yes, I have noticed this is several places with structured matrices. Would you be ok with a warning instead of an error (`Bidiag` times `Tridag` runs before this patch, having it error would be breaking)?