@andreasnoack 2019-08-07
I.e. dense output with a warning?