@mcognetta 2019-08-07
I had thought Sparse output with a warning (which I guess is still breaking, but better than interrupting execution).