@andreasnoack 2019-08-09
I guess we could also continue to return a dense array?