@dkarrasch 2019-10-23
I kind of agree that, if the result type is none of `LinearAlgebra`'s, the result should be dense. So if someone wants to invoke the sparse machinery, then (s)he would need to make at least one of the factors sparse. I remember I optimized the sparse constructors for structured matrices over at #32492, so hopefully that shouldn't be a computational bottleneck.

But, I need to admit that I'm confused. In a fresh Julia session:
```julia
using LinearAlgebra
B = Bidiagonal(rand(4), rand(3), 'U')
@which similar(B, Float64, (3,3))
```
returns
```julia
similar(B::Bidiagonal, ::Type{T}, dims::Union{Tuple{Int64}, Tuple{Int64,Int64}}) where T in SparseArrays at /Users/julia/buildbot/worker/package_macos64/build/usr/share/julia/stdlib/v1.3/SparseArrays/src/SparseArrays.jl:50
```
I couldn't find where we import this `similar` method from `SparseArrays` in `LinearAlgebra`. Shouldn't it throw a `MethodError`? I was digging a bit, but couldn't find any explicit reference to `sparse` or `Sparse` in `LinearAlgebra`, so I assume the only dependency is via these `similar` methods, right? Also, in the `Project.toml` of `LinearAlgebra`, `SparseArrays` occurs only in the extras and for testing.