@mcognetta 2020-03-21
Yeah this sort of stalled. Happy to keep working on it, perhaps with @ranocha if there is interest.