@dkarrasch 2021-05-04
`BiTriSym * BiTriSym` is currently governed by https://github.com/JuliaLang/julia/blob/ab04170ece55886d624c9836140693be2426cfa9/stdlib/SparseArrays/src/SparseArrays.jl#L58-L61

I don't understand how this is "known" to Julia without declaring `using SparseArrays`. Anyway, this PR here doesn't change the sparse return type, it only removes the case of different (in the sense of uplo) bidiagonal matrices from that list and makes it return a tridiagonal. So, we could (i) go with it, but remove the tests for `SparseMatrix` return types, or (ii) remove that special casing, and only go with the `*(::StridedMatrix, ::BiTriSym)` part. Personally, I think adding the bidiag-bidiag special case is much less important than disentangling the stdlibs,so I'd favor option (ii).