@KristofferC 2021-05-04
> I don't understand how this is "known" to Julia without declaring using SparseArrays. 

Type piracy in combination with that SparseArrays exist in the sysimage.