@mcognetta 2021-10-10
Seems all of these optimizations have been added in other PRs so I will close this one.